Added The two primary versions of DiceRoller
These are the two primary versions of a program that "rolls" dice. The 3.3.0 version is the latest and best, as it has a relatively good color scheme and lets you select the amount you want to add to your roll. The 2.1.2 version is very lightweight (not that it really matters) and also is just easier to handle because it is about half the size of the 3.3.0 version.